ksmbd: smb1: implement SMB_COM_QUERY_INFORMATION_DISK command
SMB_COM_QUERY_INFORMATION_DISK is marked as deprecated, but smbclient
will use it if SMB Trans2/QUERY_FS_INFO fails. ksmbd will disconnect
the client because the command is not implemented.

The response to this command contains the same information as
QUERY_FS_SIZE_INFO/QUERY_FS_FULL_SIZE_INFO infolevels. The difference is
that the fields are u16, so they may not be large enough. Values are
adjusted so that the client can determine free and used space in bytes
(which is what it really cares about).

Signed-off-by: Marios Makassikis <mmakassikis@freebox.fr>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>